Move cqc_changes files into own folder and separate out functions
