+++ package.json
@@ -12,6 +12,8 @@
   "dependencies": {
     "await-exec": "^0.1.2",
     "aws-sdk": "^2.771.0",
+    "axios": "^0.21.4",
+    "axios-retry": "^3.2.0",
     "bluebird": "^3.7.2",
     "bree": "^3.3.0",
     "convict": "^6.0.0",
@@ -20,8 +22,10 @@
     "humanize-duration": "^3.24.0",
     "js-yaml": "^3.14.1",
     "knex": "^0.21.6",
+    "p-ratelimit": "^1.0.1",
     "pg": "^8.4.0",
-    "pg-query-stream": "^3.3.1"
+    "pg-query-stream": "^3.3.1",
+    "sequelize": "^6.6.5"
   },
   "devDependencies": {
     "eslint": "^7.11.0",
